Refactoring. Classes moved and cleaned up.
