--- model/src/main/java/jm/model/message/ThreadChannelMessage.java
@@ -1,19 +1,0 @@
-//package jm.model.message;
-//
-//import lombok.*;
-//
-//import javax.persistence.Entity;
-//import javax.persistence.ManyToOne;
-//import javax.persistence.Table;
-//
-//@Getter
-//@Setter
-//@NoArgsConstructor
-//@EqualsAndHashCode(onlyExplicitlyIncluded = true, callSuper = true)
-//@ToString
-//@Entity
-//@Table(name = "thread_channel_messages")
-//public class ThreadChannelMessage extends Message {
-//    @ManyToOne
-//    private Message parentMessage;
-//}
