--- model/src/main/java/jm/model/message/ThreadDirectMessage.java
@@ -1,17 +1,0 @@
-//package jm.model.message;
-//
-//import lombok.*;
-//
-//import javax.persistence.*;
-//
-//@Getter
-//@Setter
-//@NoArgsConstructor
-//@EqualsAndHashCode(onlyExplicitlyIncluded = true, callSuper = true)
-//@ToString
-//@Entity
-//@Table(name = "thread_direct_messages")
-//public class ThreadDirectMessage extends Message {
-//    @ManyToOne
-//    private DirectMessage parentDirectMessage;
-//}
