+++ service-abstract/src/main/java/jm/MessageService.java
@@ -1,6 +1,6 @@
 package jm;
 
-import jm.model.message.Message;
+import jm.model.Message;
 
 import java.time.LocalDateTime;
 import java.util.List;
