+++ webapp/src/main/java/jm/controller/rest/BotRestController.java
@@ -9,7 +9,7 @@ import jm.dto.BotDtoService;
 import jm.model.Bot;
 import jm.model.Channel;
 import jm.model.Workspace;
-import jm.model.message.Message;
+import jm.model.Message;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
