+++ webapp/src/main/java/jm/controller/rest/MessageRestController.java
@@ -3,7 +3,7 @@ package jm.controller.rest;
 import jm.MessageService;
 import jm.dto.MessageDTO;
 import jm.dto.MessageDtoService;
-import jm.model.message.Message;
+import jm.model.Message;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
